Adding comment about Kubernetes dashboard
Tim suggested that we point out that the Kubernetes dashboard is moving from stable to a beta out of necessity, but this isn't called out anywhere. He suggested an addition in the release notes to point out that it is known to be a beta product and users should expect beta behavior.