Updated the kubectl run command example as it has been long deprecated
Hi all, 

The kubectl run command examples which are used on the integrator pages are long-since deprecated. 
I've updated the example to include a working command for kubectl create deployment as opposed to kubectl run. 
This was deprecated as of 1.17 if I recall, but is completely broke as of 1.18+. 

Cheers, 

- Calvin